Flor Updated

+++ Sprint-5/Tasca 1/S05T01N02SansMollLuisa/src/main/java/cat/itacademy/barcelonactiva/sansmoll/luisa/s05/t01/n02/model/services/FlorServiceImpl.java
@@ -2,6 +2,8 @@ package cat.itacademy.barcelonactiva.sansmoll.luisa.s05.t01.n02.model.services;
 
 import cat.itacademy.barcelonactiva.sansmoll.luisa.s05.t01.n02.model.domain.Flor;
 import cat.itacademy.barcelonactiva.sansmoll.luisa.s05.t01.n02.model.dto.FlorDTO;
+import cat.itacademy.barcelonactiva.sansmoll.luisa.s05.t01.n02.model.exceptions.FlorAlreadyExists;
+import cat.itacademy.barcelonactiva.sansmoll.luisa.s05.t01.n02.model.exceptions.FlorNotFound;
 import cat.itacademy.barcelonactiva.sansmoll.luisa.s05.t01.n02.model.repository.FlorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,11 +18,10 @@ public class FlorServiceImpl implements FlorService {
     @Autowired
     private FlorRepository florRepository;
 
-
     @Override
     public FlorDTO addFlor(FlorDTO florDTO) {
         if(florRepository.existsByNombreFlorAndPaisFlor(florDTO.getNombreFlor(), florDTO.getPaisFlor())){
-            throw new RuntimeException("Ya existe una flor con el mismo nombre y país.");
+            throw new FlorAlreadyExists("Ya existe una flor con el mismo nombre y país.");
 
         } else {
             Flor flor = DTOtoEntity(florDTO);
@@ -31,28 +32,44 @@ public class FlorServiceImpl implements FlorService {
 
     @Override
     public FlorDTO getOneFlor(int idFlor) {
-        Flor florbuscada = florRepository.findById(idFlor).orElse(null);
-        return (florbuscada != null) ? EntitytoDTO(florbuscada) : null;
+        Flor florBuscada = florRepository.findById(idFlor).orElse(null);
+
+        if(florBuscada == null){
+            throw new FlorNotFound("No se ha encontrado la flor con el ID "+ idFlor);
+        } else {
+            return EntitytoDTO(florBuscada);
+        }
     }
 
     @Override
     public FlorDTO getOneFlor(String nombreFlor) {
         Flor florBuscada = florRepository.findByNombreFlor(nombreFlor);
-        return (florBuscada != null) ? EntitytoDTO(florBuscada) : null;
+
+        if(florBuscada == null){
+            throw new FlorNotFound("No se ha encontrado la flor con el nombre "+ nombreFlor);
+        } else {
+            return EntitytoDTO(florBuscada);
+        }
     }
 
     @Override
     public List<FlorDTO> getAllFlores() {
         List<Flor> flores = florRepository.findAll();
-        return flores.stream()
-                .map(FlorServiceImpl::EntitytoDTO)
-                .collect(Collectors.toList());
+
+        if(flores.isEmpty()){
+            throw new FlorNotFound("No se han encontrado flores, la lista esta vacía.");
+
+        } else {
+            return flores.stream()
+                    .map(FlorServiceImpl::EntitytoDTO)
+                    .collect(Collectors.toList());
+        }
     }
 
     @Override
-    public FlorDTO updateFlor(int id, FlorDTO florDTO) {
+    public FlorDTO updateFlor(int idFlor, FlorDTO florDTO) {
         Flor nuevaFlor = DTOtoEntity(florDTO);
-        Optional<Flor> florbuscada = florRepository.findById(id);
+        Optional<Flor> florbuscada = florRepository.findById(idFlor);
 
         if(florbuscada.isPresent()){
             Flor flor = florbuscada.get();
@@ -62,18 +79,30 @@ public class FlorServiceImpl implements FlorService {
             return EntitytoDTO(flor);
 
         } else {
-            return null;
+            throw new FlorNotFound("No se ha encontrado la flor con el ID "+ idFlor);
         }
     }
 
     @Override
     public void deleteById(int idFlor) {
-        florRepository.deleteById(idFlor);
+        Optional<Flor> florbuscada = florRepository.findById(idFlor);
+
+        if(florbuscada.isPresent()){
+            florRepository.deleteById(idFlor);
+        } else {
+            throw new FlorNotFound("No se ha encontrado la flor con el ID "+ idFlor);
+        }
     }
 
     @Override
     public void deleteByNombre(String nombreFlor) {
-        florRepository.deleteByNombreFlor(nombreFlor);
+        Flor florbuscada = florRepository.findByNombreFlor(nombreFlor);
+
+        if(florbuscada == null){
+            florRepository.deleteByNombreFlor(nombreFlor);
+        } else {
+            throw new FlorNotFound("No se ha encontrado la flor con el nombre "+ nombreFlor);
+        }
     }
 
     private static Flor DTOtoEntity(FlorDTO florDTO) {
